+++ xdef/changelog.md
@@ -4,7 +4,7 @@
 * .
 
 # Version 42.1.5, release-date 2024-05-20
-* It in reported an error if the name of type in the declaration section is same
+* It is reported an error if the name of type in the declaration section is same
   as the referred type.
 * All declared user types in the conversion of X-definition to XML schema are
   written to separated XML schema file.