+++ xdef/changelog.md
@@ -3,6 +3,8 @@
 # Version 42.1.6, release-date 2024-05-24
 * Fixed bug that did not report XDEF385 error when type name in local
   declaration section was the same as the referenced type.
+* Fixed bug generated negative value to `maxOccurs` attribute in 
+  X-definition conversion to XML Schema.
 
 # Version 42.1.5, release-date 2024-05-20
 * If the type name in the declaration section is the same as the referenced type